Refactor duration out into a store
