+++ src/lib/stores.js
@@ -5,7 +5,7 @@ import { writable } from "svelte/store";
  * @property {number} id - The ID of the task
  * @property {string} category - The category of the task
  * @property {number} timeSpent - Time spent on the task in seconds
- * @property {string} startTime - The start timestamp of the task  // TODO rename to startTime
+ * @property {string} startTime - The start timestamp of the task
  * @property {string} endTime - The end timestamp of the task
  * @property {string} status - The status of the task, either 'ACTIVE' or 'COMPLETE'
  */
@@ -23,3 +23,8 @@ export let times = writable({
     tasks: [],
     error: ''
 });
+
+/**
+ * @type {import('svelte/store').Writable<number>}
+ */
+export let duration = writable(0);
